refactor: Switch coordinates to i32 to simplify calculations

+++ src/character.rs
@@ -1,23 +1,23 @@
-const SPRITE_WIDTH: usize = 60;
-const SPRITE_HEIGHT: usize = 60;
-const NUM_SPRITES: usize = 6;
+const SPRITE_WIDTH: i32 = 60;
+const SPRITE_HEIGHT: i32 = 60;
+const NUM_SPRITES: i32 = 6;
 
-const BASE_LEFT: usize = 17;
-const BASE_RIGHT: usize = 39;
-const BASE_Y: usize = 49;
+const BASE_LEFT: i32 = 17;
+const BASE_RIGHT: i32 = 39;
+const BASE_Y: i32 = 49;
 
 use crate::image::Image;
 
 /// Character struct containing position and sprite state
 pub struct Character {
-    pub dest_x: usize,
-    pub dest_y: usize,
-    pub sprite_index: usize,
+    pub dest_x: i32,
+    pub dest_y: i32,
+    pub sprite_index: i32,
 }
 
 impl Character {
     /// Create a new Character instance at the center of the screen
-    pub fn new(screen_width: usize, screen_height: usize) -> Character {
+    pub fn new(screen_width: i32, screen_height: i32) -> Character {
         Character {
             dest_x: screen_width / 2,
             dest_y: screen_height / 2,
@@ -33,7 +33,7 @@ impl Character {
             self.dest_y += 1;
         }
 
-        self.sprite_index = ((time / 200.0) as usize) % NUM_SPRITES;
+        self.sprite_index = (time / 200.0) as i32 % NUM_SPRITES;
     }
 
     fn is_on_ground(&self, screen: &Image) -> bool {
@@ -47,15 +47,15 @@ impl Character {
 
     /// Draw the character sprite to the pixel buffer
     pub fn draw(&self, screen: &mut Image, sprite_sheet: &Image) {
-        let sprite_x = (self.sprite_index * SPRITE_WIDTH) as usize;
+        let sprite_x = self.sprite_index * SPRITE_WIDTH;
         screen.draw(
             self.dest_x,
             self.dest_y,
             sprite_sheet,
             sprite_x,
             0,
-            SPRITE_WIDTH as usize,
-            SPRITE_HEIGHT as usize,
+            SPRITE_WIDTH,
+            SPRITE_HEIGHT,
         );
     }
 }
